Create URL_Fetch.gs as a simple Google Script
This code was inspired by famous "cURL" and "wget" methods to grab remote files in unix/linux systems. I wanted to create a way to do this with Google Scripts to enable the same thing and contain it on the web. This script can be fired off by hardcoding the username, password, url, and output_file_name.