Optimize Script
Remove unnecessary checks switch &&s to logical equivalent ||s and duplicates cannot occur. The Function is still valuable and I don't know where to put it. So, I wrote a function prologue for it that could be used if I ever have a need for it again. I eliminated all the fluff. The comments were removed at this point as well due to redundancy.